refactor: split listen_for_node_add tests into focused submodules
